fix: modify notification stystem in components

+++ apps/ease-up-api/src/components/board-article/board-article.service.ts
@@ -139,27 +139,32 @@ export class BoardArticleService {
 			modifier: modifier,
 		});
 		if (!result) throw new InternalServerErrorException(Message.SOMETHING_WENT_WRONG);
+		const article = await this.getBoardArticle(null, likeRefId);
+		const liker = await this.memberService.getMember(null, memberId);
 		const receiverId = await this.getMemberId(input.likeRefId);
-		const notification: NotificationInput = {
-			notificationType: NotificationType.LIKE,
-			notificationGroup: NotificationGroup.ARTICLE,
-			notificationTitle: 'Someone liked your article',
-			authorId: memberId,
-			receiverId: receiverId,
-			productId: null,
-			articleId: likeRefId,
-		};
-		if (modifier === 1) {
-			await this.notificationService.createNotification(notification);
-		} else {
-			const input = {
+
+		if (memberId.toString() !== receiverId.toString()) {
+			const notification: NotificationInput = {
+				notificationType: NotificationType.LIKE,
+				notificationGroup: NotificationGroup.ARTICLE,
+				notificationTitle: `${liker.memberNick} has liked your article '${article.articleTitle}'`,
 				authorId: memberId,
 				receiverId: receiverId,
+				productId: null,
 				articleId: likeRefId,
 			};
-			await this.notificationService.deleteNotification(input);
+			if (modifier === 1) {
+				await this.notificationService.createNotification(notification);
+			} else {
+				const input = {
+					authorId: memberId,
+					receiverId: receiverId,
+					articleId: likeRefId,
+				};
+				await this.notificationService.deleteNotification(input);
+			}
+			return result;
 		}
-		return result;
 	}
 
 	public async getMemberId(articleId: ObjectId): Promise<ObjectId> {
