+++ apps/ease-up-api/src/components/property/property.service.ts
@@ -219,25 +219,30 @@ export class PropertyService {
 			modifier: modifier,
 		});
 		if (!result) throw new InternalServerErrorException(Message.SOMETHING_WENT_WRONG);
+
+		const property = await this.getProperty(null, likeRefId);
+		const liker = await this.memberService.getMember(null, memberId);
 		const receiverId = await this.getMemberId(input.likeRefId);
-		const notification: NotificationInput = {
-			notificationType: NotificationType.LIKE,
-			notificationGroup: NotificationGroup.PROPERTY,
-			notificationTitle: 'Someone liked your product',
-			authorId: memberId,
-			receiverId: receiverId,
-			productId: likeRefId,
-			articleId: null,
-		};
-		if (modifier === 1) {
-			await this.notificationService.createNotification(notification);
-		} else {
-			const input = {
+		if (memberId.toString() !== receiverId.toString()) {
+			const notification: NotificationInput = {
+				notificationType: NotificationType.LIKE,
+				notificationGroup: NotificationGroup.PROPERTY,
+				notificationTitle: `${liker.memberNick} has liked your product '${property.propertyTitle}'`,
 				authorId: memberId,
 				receiverId: receiverId,
 				productId: likeRefId,
+				articleId: null,
 			};
-			await this.notificationService.deleteNotification(input);
+			if (modifier === 1) {
+				await this.notificationService.createNotification(notification);
+			} else {
+				const input = {
+					authorId: memberId,
+					receiverId: receiverId,
+					productId: likeRefId,
+				};
+				await this.notificationService.deleteNotification(input);
+			}
 		}
 		return result;
 	}
